+++ recusal/__init__.py
@@ -67,4 +67,4 @@ __all__ = [
     "DEFAULT_TAXONOMY",
 ]
 
-__version__ = "0.5.8"
+__version__ = "0.5.9"